Rename hskip.js -> kern.js
